[refactor] Move extensions to single file
