+++ Features/Tests/FeaturesTests/FeaturesTests.swift
@@ -36,17 +36,3 @@ final class AppFeaturesTests: XCTestCase {
     }
   }
 }
-
-extension XCTestCase {
-  func getMockMultipleQuestions() -> [String] {
-    [
-      "身上使用最久的東西是什麼？",
-      "最喜歡的一部電影？",
-      "今年最期待的一件事情？",
-      "我不為人知的一個奇怪技能",
-      "做過最像大人的事情",
-      "今年最快樂的回憶",
-      "最想再去一次的國家/城市",
-    ]
-  }
-}
